refactor(provider): move ProviderRO and ProviderRW traits to katana-provider
