+++ crates/rpc/rpc-server/src/starknet/pending.rs
@@ -1,6 +1,5 @@
 use std::fmt::Debug;
 
-use katana_core::backend::storage::ProviderRO;
 use katana_core::service::block_producer::{BlockProducer, BlockProducerMode};
 use katana_executor::ExecutorFactory;
 use katana_primitives::block::PartialHeader;
@@ -9,7 +8,7 @@ use katana_primitives::execution::TypedTransactionExecutionInfo;
 use katana_primitives::transaction::{TxHash, TxNumber};
 use katana_primitives::version::CURRENT_STARKNET_VERSION;
 use katana_provider::api::state::StateProvider;
-use katana_provider::ProviderFactory;
+use katana_provider::{ProviderFactory, ProviderRO};
 use katana_rpc_types::{
     FinalityStatus, PreConfirmedBlockWithReceipts, PreConfirmedBlockWithTxHashes,
     PreConfirmedBlockWithTxs, PreConfirmedStateUpdate, ReceiptBlockInfo, RpcTxWithHash,
