+++ crates/utils/src/node.rs
@@ -2,7 +2,6 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 
 use katana_chain_spec::{dev, ChainSpec};
-use katana_core::backend::storage::{ProviderRO, ProviderRW};
 use katana_core::backend::Backend;
 use katana_executor::implementation::blockifier::BlockifierFactory;
 use katana_node::config::dev::DevConfig;
@@ -12,7 +11,9 @@ use katana_node::config::Config;
 use katana_node::{LaunchedNode, Node};
 use katana_primitives::address;
 use katana_primitives::chain::ChainId;
-use katana_provider::{DbProviderFactory, ForkProviderFactory, ProviderFactory};
+use katana_provider::{
+    DbProviderFactory, ForkProviderFactory, ProviderFactory, ProviderRO, ProviderRW,
+};
 use katana_rpc_server::HttpClient;
 use starknet::accounts::{ExecutionEncoding, SingleOwnerAccount};
 use starknet::core::types::BlockTag;
